pyproject.toml
[tool.poetry]
name = "ai-video-cli"
version = "0.1.0"
description = "AI Video Editor CLI Tool"
authors = ["Marcus Schiesser <mail@marcusschiesser.de>"]
readme = "README.md"
packages = [{ include = "ai_video_cli" }]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
moviepy = "^1.0.3"
numpy = "^1.26.0"
pillow = "^9"

[tool.poetry.scripts]
ai-video = "ai_video_cli.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
